+++ packages/ui/spinner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nsui/spinner
 
+## 1.0.1
+
+### Patch Changes
+
+- Separate scale transition from other transitions, now managed by `colorTransitionDuration` and `scaleTransitionDuration` (previously `transitionDuration`)
+- Updated dependencies
+  - @nsui/provider@1.4.0
+
 ## 1.0.0
 
 ### Major Changes